+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.2.0] - 2022/05/10
 ### Added
 - New modal - `Local Quote Statistics` with basic local quotes' stats
 - Ability to inherit listing's style in quote block (enable via settings, off by default)